requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-generativeai
PyPDF2
python-docx